# native/internal/openapi/swagger.yaml
basePath: /
definitions:
...
  github_com_gcc798_microservice-kit_application_iam_internal_response.Response:
    description: API 统一响应格式
    properties:
      code:
        description: 业务状态码
        example: 200
        type: integer
      data:
        description: 响应数据
        type: object
      msg:
        description: 响应消息
        example: 操作成功
        type: string
    type: object
...
  model.Role:
    properties:
      createBy:
        description: 创建人
        type: integer
      createdTime:
        type: string
      dataScope:
        description: 数据范围：1全部 2自定义 3本组织 4本组织及以下 5仅本人
        type: integer
      id:
        description: 角色ID（使用分布式ID）
        type: integer
      isSystem:
        description: 是否系统内置角色（内置角色不可删除）
        type: boolean
      remark:
        description: 备注
        type: string
      roleKey:
        description: 角色标识（唯一，用于权限匹配）
        type: string
      roleName:
        description: 角色名称
        type: string
      sort:
        description: 显示顺序
        type: integer
      status:
        description: 状态：0正常 1停用
        type: integer
      updateBy:
        description: 更新人
        type: integer
      updatedTime:
        type: string
    type: object
  request.AssignRoleToUserRequest:
    properties:
      roleId:
        description: 角色ID
        example: 1
        type: integer
      userId:
        description: 用户ID
        example: 1001
        type: integer
    required:
    - roleId
    - userId
    type: object
...
  request.CreateRoleRequest:
    properties:
      dataScope:
        description: 数据范围：1全部 2自定义 3本组织 4本组织及以下 5仅本人
        example: 2
        type: integer
      remark:
        description: 备注
        example: 负责用户管理
        type: string
      roleKey:
        description: 角色标识（唯一）
        example: user_manager
        type: string
      roleName:
        description: 角色名称
        example: 用户管理员
        type: string
      sort:
        description: 显示顺序
        example: 1
        type: integer
      status:
        description: 状态：0正常 1停用
        example: 0
        type: integer
    required:
    - roleKey
    - roleName
    type: object
...
  request.PageRoleRequest:
    properties:
      isAsc:
        description: 'IsAsc 排序方向，asc或desc，支持多个，逗号分隔，如: "asc,desc"'
        type: string
      orderByColumn:
        description: 'OrderByColumn 排序列，支持多列排序，逗号分隔，如: "id,createTime"'
        type: string
      pageNum:
        description: PageNum 当前页码，从1开始
        type: integer
      pageSize:
        description: PageSize 每页数量
        type: integer
      roleName:
        description: 角色名称（模糊查询）
        example: 管理员
        type: string
      status:
        description: 状态：0正常 1停用，-1表示不过滤
        example: 0
        type: integer
    type: object
...
  request.UpdateRoleRequest:
    properties:
      dataScope:
        description: 数据范围
        example: 2
        type: integer
      remark:
        description: 备注
        example: 负责用户管理
        type: string
      roleId:
        description: 角色ID
        example: 1
        type: integer
      roleName:
        description: 角色名称
        example: 用户管理员
        type: string
      sort:
        description: 显示顺序
        example: 1
        type: integer
      status:
        description: 状态：0正常 1停用
        example: 0
        type: integer
    required:
    - roleId
    - roleName
    type: object
...
host: localhost:9009
info:
  contact: {}
  description: microservice-kit RESTful API
  title: microservice-kit API
  version: "1.0"
paths:
...
  /api/v1/role:
    post:
      consumes:
      - application/json
      description: 创建新角色
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Role'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      summary: 创建角色
      tags:
      - 角色管理
  /api/v1/role/{roleId}:
    delete:
      consumes:
      - application/json
      description: 删除角色
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 角色ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      summary: 删除角色
      tags:
      - 角色管理
    get:
      consumes:
      - application/json
      description: 根据角色ID获取角色详情
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 角色ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Role'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      summary: 获取角色详情
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 更新角色信息
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 角色ID
        in: path
        name: roleId
        required: true
        type: integer
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      summary: 更新角色
      tags:
      - 角色管理
  /api/v1/role/assign:
    post:
      consumes:
      - application/json
      description: 为用户分配角色
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分配信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AssignRoleToUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      summary: 为用户分配角色
      tags:
      - 角色管理
  /api/v1/role/page:
    post:
      consumes:
      - application/json
      description: 使用 Paginator 分页查询角色列表
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.PageRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: 分页查询角色列表
      tags:
      - 角色管理
  /api/v1/role/remove:
    delete:
      consumes:
      - application/json
      description: 移除指定用户的角色
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: query
        name: userId
        required: true
        type: integer
      - description: 角色ID
        in: query
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      summary: 移除用户的角色
      tags:
      - 角色管理
  /api/v1/role/user:
    get:
      consumes:
      - application/json
      description: 获取用户的所有角色
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: query
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      summary: 获取用户的所有角色
      tags:
      - 角色管理
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"